requirements.txt
streamlit
pdfminer.six
ebooklib
pillow
pdf2image
